test fixtures reorganization WIP
